requirements.txt
Flask
requests
slack_sdk
python-dotenv
